add server code for a RESTful API & an admin panel
This is the very first commit of the 'portfolio-backend' repository. It is a Strapi app, written in Typescript, running on NodeJS, that will serve a RESTful api (mainly for CRUD operations) and an admin web panel. The project was scaffolded by running `yarn create strapi-app portfolio-backend --typescript --no-run --quickstart` and all generated files were initially commited (added to the repository).